backend/requirements.txt
fastapi
uvicorn
pymongo
bcrypt
pyjwt
langchain
langchain-community
langchain-huggingface
openai
python-multipart